Separate cube rendering
